requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
mysql-connector-python==9.0.0
pydantic==2.5.2
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1